+++ validator-api/Cargo.toml
@@ -7,7 +7,7 @@ version = "0.11.0"
 authors = [
     "Dave Hrycyszyn <futurechimp@users.noreply.github.com>",
     "Jędrzej Stuczyński <andrew@nymtech.net>",
-    "Drazen Urch",
+    "Drazen Urch <durch@users.noreply.github.com>" 
 ]
 edition = "2018"
